[SPARK-2320] Reduce <pre> block font size.
